% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simScenario5.R
\docType{data}
\name{simScenario5}
\alias{simScenario5}
\title{Simulated data from simulation scenario 5}
\format{
A data frame with 400 obs. and 8 variables.
}
\usage{
data(simScenario5)
}
\description{
Simulated data in the supplement to Follmann, et al (2020). T1,T2, and T3
are the time to the first event for three different types of events
(e.g., Death, Stroke/MI, Bleed). I1,I2, and I3 are the associated
status variables (0=censored, 1=event). Other variables are
PATID (patient ID) and Z (0=control arm, 1=treatment arm).
}
\references{
Follmann, D., Fay, M. P., Hamasaki, T., and Evans, S. (2020). Analysis of
ordered composite endpoints. Statistics in Medicine, 39(5), 602-616.
}
\keyword{datasets}
